requirements.txt
Django==1.8.3
gunicorn==19.3.0
psycopg2==2.6.1
Pillow==2.9.0
django-storages==1.1.8
boto==2.38.0
django-imagekit==3.2.6
